% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{api_token}
\alias{api_token}
\title{Retrieve Qualtrics API token}
\usage{
api_token(.f = "~/.bpqx-auth")
}
\arguments{
\item{.f}{String pointing to the location of the credentials file}
}
\description{
Retrieve Qualtrics API token
}
