% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_responses.R
\name{fetch_responses}
\alias{fetch_responses}
\title{Fetch responses for a given survey}
\usage{
fetch_responses(survey_id)
}
\arguments{
\item{survey_id}{String: a unique identifier for the target survey}
}
\description{
Fetch responses for a given survey
}
