% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_cache.R
\name{qx_cache_path}
\alias{qx_cache_path}
\title{Produce a path in the bpqx data cache}
\usage{
qx_cache_path(..., .p = NULL)
}
\arguments{
\item{...}{additional path nodes}

\item{.p}{cache root, automatically configured by use_bpqx}
}
\description{
Produce a path in the bpqx data cache
}
