% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_cache.R
\name{qx_cache_root}
\alias{qx_cache_root}
\title{Produce the path to the root folder of the qx_cache.
Prompt setup with use_bpqx if missing.
Prompt vpn connection if set up but doesn't exist.}
\usage{
qx_cache_root()
}
\description{
Produce the path to the root folder of the qx_cache.
Prompt setup with use_bpqx if missing.
Prompt vpn connection if set up but doesn't exist.
}
