% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx_request.R
\name{qx_req}
\alias{qx_req}
\title{Create a request url for httr2::request}
\usage{
qx_req(...)
}
\arguments{
\item{...}{strings, beginning with the endpoint,
representing the additional pathing following ...API/v3}
}
\description{
Create a request url for httr2::request
}
