% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{save_api_token}
\alias{save_api_token}
\title{Stores Qualtrics API token in "~/.bpqx-auth"}
\usage{
save_api_token(
  data_center = "blueprintade.yul1",
  .t = NULL,
  .f = "~/.bpqx-auth"
)
}
\arguments{
\item{data_center}{A string specifying the data center to use}

\item{.t}{A string which overrides interactive prompt.
strictly for automated testing.}
}
\description{
Stores Qualtrics API token in "~/.bpqx-auth"
}
